4 -> 64

+++ app/protocol.rs
@@ -124,7 +124,7 @@ impl From<Command> for List {
                 Integer(2),
                 cons(
                     Integer(ship_id),
-                    cons(cons(Integer(x), Integer(y)), cons(Integer(4), Nil)),
+                    cons(cons(Integer(x), Integer(y)), cons(Integer(64), Nil)),
                 ),
             ),
         }
